Finished overhaul
