+++ launcher/ui/__init__.py
@@ -5,10 +5,20 @@ __all__ = [
     "NewsTab",
     "ServerStatusBar",
     "Background",
+    "TopBar",
+    "Logo",
+    "ProgressBar",
+    "InstallationDialog",
+    "Button",
 ]
 
 from launcher.ui.background import Background
+from launcher.ui.button import Button
+from launcher.ui.installation_dialog import InstallationDialog
 from launcher.ui.login import Login
+from launcher.ui.logo import Logo
 from launcher.ui.news_area import NewsArea, NewsTab
+from launcher.ui.progress_bar import ProgressBar
 from launcher.ui.server_status_bar import ServerStatusBar
 from launcher.ui.settings import SettingsDialog
+from launcher.ui.top_bar import TopBar
